added some type checks for functions

+++ controllers/posts.js
@@ -10,8 +10,7 @@ function create_post(req, res, next) {
     return utils.response(req, res, 400, {error: 'Missing required fields'});
   }
   
-  if (typeof (req.body.auth_user_id) != "string" || 
-      typeof (req.body.title) != "string" ||
+  if (typeof (req.body.title) != "string" ||
       typeof (req.body.content) != "string" ||
       Array.isArray(req.body.images) == false ||
       Array.isArray(req.body.keywords) == false) {
@@ -66,6 +65,7 @@ function create_post(req, res, next) {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
+    keywords: req.body.keywords,
     images: req.body.images,
     posterId: req.body.auth_user_id,
   });
@@ -81,7 +81,14 @@ function create_post(req, res, next) {
 }
 //get information about the posts, gets the post id
 function get_post(req, res, next) {
-  const postId = req.params.id;
+  const postId = req.params._id;
+
+  var ObjectId = require('mongoose').Types.ObjectId;
+  if (ObjectId.isValid(postId) == false){
+    return res.status(400).json({
+      error: 'Wrong get Post ID'
+    });
+  }
 
   Post.findOne({ _id: postId }, (err, post) => {
     if (err) {
@@ -98,6 +105,11 @@ function get_post(req, res, next) {
 
 function search_post(req, res, next) {
   var searchTerm = req.query.searchterm;
+
+  if (typeof(searchTerm) != 'string'){
+    return utils.response(req, res, 400, {error: 'Wrong searchterm type'});
+  }
+  
   var searchRegex = new RegExp('.*' + searchTerm + ".*"); //searches for any string
   const skipCount = req.query.count || 0;
 
@@ -149,6 +161,10 @@ function get_post_by_userId(req, res, next) {
   const userId = req.params.id;
   const skipCount = req.query.count;
   
+  if (typeof(skipCount) != 'number'){
+    return utils.response(req, res, 400, {error: 'Wrong skipCount type'});
+  }
+  
   var ObjectId = require('mongoose').Types.ObjectId;
   if (ObjectId.isValid(userId) == false){
     return res.status(400).json({
@@ -170,6 +186,13 @@ function get_post_by_userId(req, res, next) {
 function delete_post(req, res, next) {
   const postId = req.params.id;
 
+  var ObjectId = require('mongoose').Types.ObjectId;
+  if (ObjectId.isValid(postId) == false){
+    return res.status(400).json({
+      error: 'Invalid Post ID'
+    });
+  }
+
   Post.deleteOne({user: req.auth_user_id, _id: postId})
     .then(() => {
       return res.status(200).json({
@@ -185,6 +208,11 @@ function delete_post(req, res, next) {
 
 function get_newest_posts(req, res, next) {
   const skipCount = req.query.count || 0;
+
+  if (typeof(skipCount) != 'number'){
+    return utils.response(req, res, 400, {error: 'Invalid skipCount type'});
+  }
+
   Post.find((err, post) => {
     if (err) {
       return utils.response(req, res, 500, utils.response(req, res, 500, {error: 'Internal server error'}));
@@ -194,13 +222,70 @@ function get_newest_posts(req, res, next) {
 }
 
 function edit_post(req, res, next) {
-  if (!utils.check_body_fields(req.body, ['title', 'body', 'content', 'images','keywords'])) {
+  if (!utils.check_body_fields(req.body, [ 'post_id', 'auth_user_id', 'title', 'content', 'images','keywords'])) {
     return utils.response(req, res, 400, {error: 'Missing required fields'});
   }
+  
+  var ObjectId = require('mongoose').Types.ObjectId;
+  if (ObjectId.isValid(req.body.post_id) == false){
+    return res.status(400).json({
+      error: 'Invalid Post ID'
+    });
+  }
 
-  User.findOneAndUpdate({ _id : req.body.auth_user_id }, { "$set": {
-    title: req.body.title, 
-    body: req.body.body, 
+  if (typeof (req.body.title) != "string" ||
+      typeof (req.body.content) != "string" ||
+      Array.isArray(req.body.images) == false ||
+      Array.isArray(req.body.keywords) == false) {
+        return utils.response(req, res, 400, {error: 'Invalid type of edit'});
+  }
+   
+  // Need 1 to 10 images
+  if (req.body.images.length < 1 || req.body.images.length > 10) {
+    return utils.response(req, res, 400, {error: 'Invalid number of images'});
+  }
+
+  // Check type of elements in images[] array
+  for (let i = 0; i < (req.body.images).length; i++) { 
+    if (typeof((req.body.images)[i]) != 'string') {
+      return utils.response(req, res, 400, {error: 'Invalid type of images'});
+    }
+  }
+
+  // Check if title is too long
+  if (req.body.title.length > 25) {
+    return utils.response(req, res, 400, {error: 'Title is too long'});
+  }
+
+  // Check if content is too long
+  if (req.body.content.length > 2000) {
+    return utils.response(req, res, 400, {error: 'Content is too long'});
+  }
+
+  // Validate the image paths
+  // TODO
+
+  // Check if there are too many keywords
+  if (req.body.keywords.length > 10) {
+    return utils.response(req, res, 400, {error: 'Too many keywords'});
+  }
+
+  // Check type of elements in keywords[] array
+  for (let i = 0; i < (req.body.keywords).length; i++) { 
+    if (typeof((req.body.keywords)[i]) != 'string') {
+      return utils.response(req, res, 400, {error: 'Invalid type of keywords'});
+    }
+  }
+
+  // Check if keywords are valid
+  for (let keyword of req.body.keywords) {
+    if (!utils.is_valid_keyword(keyword)) {
+      return utils.response(req, res, 400, {error: 'Invalid keywords'});
+    }
+  }
+
+  Post.findOneAndUpdate({ _id : req.body.post_id }, { "$set": {
+    title: req.body.title,  
     content: req.body.content,
     images: req.body.images,
     keywords: req.body.keywords
@@ -212,8 +297,16 @@ function edit_post(req, res, next) {
     }
  });
 }
+
 function delete_post_image(req, res, next){
   const imageId = req.params.id;
+
+  var ObjectId = require('mongoose').Types.ObjectId;
+  if (ObjectId.isValid(imageId) == false){
+    return res.status(400).json({
+      error: 'Invalid Image ID'
+    });
+  }
 
   Image.deleteOne({_id: imageId})
     .then(() => {
